requirements.txt
psycopg2
Flask
Flask-SQLAlchemy
Flask-Login
SQLAlchemy
python-dotenv
Pillow
Werkzeug
